requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pymongo==4.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.3
email-validator==2.2.0
pytest==8.3.4
httpx==0.28.1
motor==3.6.0
